+++ nnt/TL/Variant+NNT.h
@@ -167,7 +167,7 @@ public:
         _v.refobj = v;
     }
     
-    explicit variant_t(char const* str, core::_bool_copy cpy, uint len = -1)
+    explicit variant_t(char const* str, core::_bool_copy cpy, usize len = -1)
     : vt(VT_PCHAR)
     {
         _rel = core::copy == cpy;
@@ -671,7 +671,7 @@ public:
         return *this;
     }
     
-    void set_str(char const* str, core::_bool_copy cpy, uint len = -1)
+    void set_str(char const* str, core::_bool_copy cpy, usize len = -1)
     {
         clear();
